chore(parser): clarify result compatibility boundaries
