+++ parser_result_collapsed_helpers_test.go
